docs(design): update quantity field docs
