+++ apps/design-land/src/app/quantity-field/quantity-field.module.ts
@@ -1,7 +1,10 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 
-import { DaffArticleModule } from '@daffodil/design';
+import {
+  DaffArticleModule,
+  DaffQuantityFieldModule,
+} from '@daffodil/design';
 
 import { DesignLandExampleViewerModule } from '../core/code-preview/container/example-viewer.module';
 import { DesignLandQuantityFieldRoutingModule } from './quantity-field-routing.module';
@@ -14,6 +17,7 @@ import { DesignLandQuantityFieldComponent } from './quantity-field.component';
   imports: [
     CommonModule,
     DaffArticleModule,
+    DaffQuantityFieldModule,
     DesignLandExampleViewerModule,
     DesignLandQuantityFieldRoutingModule,
   ],
