+++ libs/design/quantity-field/examples/ng-package.json
@@ -1,9 +1,9 @@
 {
-  "$schema": "../../../../node_modules/ng-packagr/ng-package.schema.json",
-  "dest": "../../../dist/design/examples",
+  "$schema": "../../node_modules/ng-packagr/ng-package.schema.json",
+  "dest": "../../dist/design/examples",
   "deleteDestPath": false,
   "lib": {
     "entryFile": "src/index.ts",
     "styleIncludePaths": ["../../src/scss"]
   }
-} 
+}
