+++ libs/design/quantity-field/examples/src/basic-quantity-field/basic-quantity-field.component.spec.ts
