+++ libs/design/quantity-field/examples/src/basic-quantity-field/basic-quantity-field.component.ts
