+++ libs/design/quantity-field/examples/src/custom-range-quantity-field/custom-range-quantity-field.component.ts
