+++ libs/design/quantity-field/examples/src/disabled-quantity-field/disabled-quantity-field.module.ts
