--- libs/design/quantity-field/examples/src/examples.ts
@@ -1,11 +1,0 @@
-import { BasicQuantityFieldComponent } from './basic/basic-quantity-field.component';
-import { CustomRangeQuantityFieldComponent } from './customRange/custom-range-quantity-field.component';
-import { DisabledQuantityFieldComponent } from './disabled/disabled-quantity-field.component';
-import { SelectMaxQuantityFieldComponent } from './selectMax/select-max-quantity-field.component';
-
-export const QUANTITY_FIELD_EXAMPLES = [
-  BasicQuantityFieldComponent,
-  CustomRangeQuantityFieldComponent,
-  DisabledQuantityFieldComponent,
-  SelectMaxQuantityFieldComponent,
-];
