+++ libs/design/quantity-field/examples/src/select-max-quantity-field/select-max-quantity-field.component.ts
